Extract prometheus/response.go
